# apiserver7dtd/testdata/openapi.snapshot.json
{
  "components": {
    "schemas": {
      "CommandResponse": {
        "properties": {
          "data": {
            "properties": {
              "command": {
                "type": "string"
              },
              "parameters": {
                "nullable": true,
                "type": "string"
              },
              "result": {
                "type": "string"
              }
            },
            "required": [
              "command",
              "parameters",
              "result"
            ],
            "type": "object"
          },
          "meta": {
            "properties": {
              "serverTime": {
                "format": "date-time",
                "nullable": true,
                "type": "string"
              }
            },
            "type": "object"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "type": "string"
              },
              "details": {
                "additionalProperties": true,
                "type": "object"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "UpdateJob": {
        "properties": {
          "backupId": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "currentVersion": {
            "type": "string"
          },
          "error": {
            "properties": {
              "code": {
                "type": "string"
              },
              "details": {
                "additionalProperties": true,
                "type": "object"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          },
          "finishedAt": {
            "format": "date-time",
            "type": "string"
          },
          "jobId": {
            "type": "string"
          },
          "phase": {
            "enum": [
              "queued",
              "checking",
              "stopping",
              "checking_stopped",
              "backing_up",
              "updating",
              "starting",
              "verifying",
              "completed"
            ],
            "type": "string",
            "x-enum-varnames": [
              "UpdateJobPhaseQueued",
              "UpdateJobPhaseChecking",
              "UpdateJobPhaseStopping",
              "UpdateJobPhaseCheckingStopped",
              "UpdateJobPhaseBackingUp",
              "UpdateJobPhaseUpdating",
              "UpdateJobPhaseStarting",
              "UpdateJobPhaseVerifying",
              "UpdateJobPhaseCompleted"
            ]
          },
          "planId": {
            "type": "string"
          },
          "recoveryRequired": {
            "type": "boolean"
          },
          "resultVersion": {
            "type": "string"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "interrupted"
            ],
            "type": "string",
            "x-enum-varnames": [
              "UpdateJobStatusQueued",
              "UpdateJobStatusRunning",
              "UpdateJobStatusSucceeded",
              "UpdateJobStatusFailed",
              "UpdateJobStatusInterrupted"
            ]
          },
          "targetVersion": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "jobId",
          "planId",
          "status",
          "phase",
          "createdAt",
          "updatedAt",
          "currentVersion",
          "targetVersion",
          "recoveryRequired"
        ],
        "type": "object"
      },
      "UpdateJobRequest": {
        "additionalProperties": false,
        "properties": {
          "confirmation": {
            "enum": [
              "UPDATE SUZUME"
            ],
            "type": "string"
          },
          "idempotencyKey": {
            "pattern": "^[A-Za-z0-9_-]{16,128}$",
            "type": "string"
          },
          "planId": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "planId",
          "confirmation",
          "idempotencyKey"
        ],
        "type": "object"
      },
      "UpdateJobResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UpdateJob"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "7 Days to Die サーバーを docker compose で操作する同期API",
    "title": "7dtd Ops API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/server/command": {
      "get": {
        "description": "7 Days to Die のゲームサーバーに管理コンソールコマンドを実行します。\n内部的には上流の `/api/command` にフォワードします。\n返却 `data.result` はコマンド出力の生文字列です。\n",
        "operationId": "execServerCommand",
        "parameters": [
          {
            "examples": {
              "say": {
                "summary": "Say something",
                "value": "say Hello from API"
              },
              "version": {
                "summary": "Show version",
                "value": "version"
              }
            },
            "in": "query",
            "name": "command",
            "schema": {
              "maxLength": 2000,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "version": {
                    "value": {
                      "data": {
                        "command": "version",
                        "parameters": "",
                        "result": "Game version: V 2.2 (b3) Compatibility Version: V 2.2\nMod TFP_Harmony: 1.1.0.4\n..."
                      },
                      "meta": {
                        "serverTime": "2025-08-18T22:33:17.875056+09:00"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            },
            "description": "実行成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "リクエスト不正"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Another operation/recovery holds the lock, or update state is unavailable"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "上流エラー"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Another operation/recovery holds the lock, or update state is unavailable"
          }
        },
        "summary": "7dtdコンソールコマンドの実行（同期）"
      }
    },
...
    "/server/update/jobs": {
      "post": {
        "description": "Requires confirmation UPDATE SUZUME and a durable idempotency key. The same request/key returns\nthe original job, even after expiry or restart. A plan can produce only one job. Fresh player\nand version checks happen asynchronously; their failures appear in job.error, after acceptance.\nNever retry an ambiguous acceptance with a different key.\n",
        "operationId": "createServerUpdateJob",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateJobResponse"
                }
              }
            },
            "description": "Same request/key; original job returned"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateJobResponse"
                }
              }
            },
            "description": "Durable job accepted; poll Location",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid confirmation, unauthorized, expired/used/blocked plan, busy, conflicting key or unavailable state"
          }
        },
        "summary": "Confirm and enqueue an asynchronous update"
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "https://example.invalid"
    }
  ]
}